requirements.txt
numpy
python-opencv
pillow
torch
transformers
tqdm
pandas
python-doctr
